feat(config): add `rmk_keyboard` macro skeleton
