+++ rmk-macro/src/keyboard_config.rs
@@ -1,0 +1,132 @@
+use quote::quote;
+use rmk_config::toml_config::{KeyboardInfo, LightConfig, MatrixConfig, PinConfig};
+
+use crate::gpio_config::{
+    convert_gpio_str_to_output_pin, convert_input_pins_to_initializers,
+    convert_output_pins_to_initializers,
+};
+
+#[derive(Debug, PartialEq, Eq)]
+pub(crate) enum ChipSeries {
+    Stm32,
+    Nrf52,
+    Rp2040,
+    Esp32,
+    Unsupported,
+}
+
+pub(crate) fn get_chip_model(chip: String) -> ChipSeries {
+    if chip.to_lowercase().starts_with("stm32") {
+        return ChipSeries::Stm32;
+    } else if chip.to_lowercase().starts_with("nrf52") {
+        return ChipSeries::Nrf52;
+    } else if chip.to_lowercase().starts_with("rp2040") {
+        return ChipSeries::Rp2040;
+    } else if chip.to_lowercase().starts_with("esp32") {
+        return ChipSeries::Esp32;
+    } else {
+        return ChipSeries::Unsupported;
+    }
+}
+
+pub(crate) fn expand_keyboard_info(keyboard_info: KeyboardInfo) -> proc_macro2::TokenStream {
+    let pid = keyboard_info.product_id;
+    let vid = keyboard_info.vendor_id;
+    let product_name = keyboard_info
+        .product_name
+        .unwrap_or("RMK Keyboard".to_string());
+    let manufacturer = keyboard_info.manufacturer.unwrap_or("RMK".to_string());
+    let serial_number = keyboard_info.serial_number.unwrap_or("0000000".to_string());
+    quote! {
+        static keyboard_usb_config: ::rmk_config::keyboard_config::KeyboardUsbConfig = ::rmk_config::keyboard_config::KeyboardUsbConfig {
+            vid: #vid,
+            pid: #pid,
+            manufacturer: #manufacturer,
+            product_name: #product_name,
+            serial_number: #serial_number,
+        };
+    }
+}
+
+pub(crate) fn expand_vial_config() -> proc_macro2::TokenStream {
+    quote! {
+        static vial_config: ::rmk_config::keyboard_config::VialConfig = ::rmk_config::keyboard_config::VialConfig {
+            vial_keyboard_id: &VIAL_KEYBOARD_ID,
+            vial_keyboard_def: &VIAL_KEYBOARD_DEF,
+        };
+    }
+}
+
+pub(crate) fn build_light_config(
+    chip: &ChipSeries,
+    pin_config: Option<PinConfig>,
+) -> proc_macro2::TokenStream {
+    match pin_config {
+        Some(c) => {
+            let p = convert_gpio_str_to_output_pin(chip, c.pin);
+            let low_active = c.low_active;
+            quote! {
+                Some(::rmk_config::keyboard_config::LightPinConfig {
+                    pin: #p,
+                    low_active: #low_active,
+                })
+            }
+        }
+        None => quote! {None},
+    }
+}
+
+pub(crate) fn expand_light_config(
+    chip: &ChipSeries,
+    light_config: LightConfig,
+) -> proc_macro2::TokenStream {
+    let numslock = build_light_config(chip, light_config.numslock);
+    let capslock = build_light_config(chip, light_config.capslock);
+    let scrolllock = build_light_config(chip, light_config.scrolllock);
+
+    // Generate a macro that does light config
+    quote! {
+        macro_rules! config_light {
+            (p: $p:ident) => {{
+                let light_config = ::rmk_config::keyboard_config::LightConfig {
+                    capslock: #capslock,
+                    numslock: #numslock,
+                    scrolllock: #scrolllock,
+                };
+                (light_config)
+            }};
+        }
+    }
+}
+
+pub(crate) fn expand_matrix_config(
+    chip: &ChipSeries,
+    matrix_config: MatrixConfig,
+) -> proc_macro2::TokenStream {
+    let num_col = matrix_config.cols as usize;
+    let num_row = matrix_config.rows as usize;
+    let num_layer = matrix_config.layers as usize;
+    let mut final_tokenstream = proc_macro2::TokenStream::new();
+    final_tokenstream.extend(convert_input_pins_to_initializers(
+        &chip,
+        matrix_config.input_pins,
+    ));
+    final_tokenstream.extend(convert_output_pins_to_initializers(
+        &chip,
+        matrix_config.output_pins,
+    ));
+
+    // Generate a macro that does pin matrix config
+    quote! {
+        pub(crate) const COL: usize = #num_col;
+        pub(crate) const ROW: usize = #num_row;
+        pub(crate) const NUM_LAYER: usize = #num_layer;
+
+        macro_rules! config_matrix {
+            (p: $p:ident) => {{
+                #final_tokenstream
+                (output_pins, input_pins)
+            }};
+        }
+    }
+}
